move search-related test methods from MetadataRepositoryTest into new SearchServiceTest class
